% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulation.R
\name{plot_simulation}
\alias{plot_simulation}
\title{Plots the SIR epidemic from the simulation}
\usage{
plot_simulation(results, num_areas, num_ages, by_age = F, step = 1,
  pretty = F, title = NULL)
}
\arguments{
\item{results}{The object containing the results from the simulation.}

\item{num_areas}{The number of areas in the simulation. Easy way to get it for your current simulation is
to use the dimension of your transmission kernel (e.g. if you have 4 age groups:
dim(expanded_D)[1]/4).}

\item{num_ages}{The number of ages in the simulation.}

\item{by_age}{Logical. If TRUE, will create a plot for each age category (only works with pretty = FALSE).}

\item{step}{Timestep used for the simulation, in days (default is 1 day).}

\item{pretty}{Logical. If TRUE, will render using ggplot (longer but nicer).}

\item{title}{Title to use for the plot.}
}
\value{
Creates a plot.
}
\description{
Plots the total fraction of Susceptible, Infected and Recovered individuals at each time step.
}
\details{
The Infected category has a separate y-axis to make visualisation easier, due to the relatively low
         fraction of Infected at any given time compared to Susceptible and Recovered.
}
\examples{

#Create a spatial dataset:
test_data = raster(nrow=10, ncol=10, xmn=1, xmx=100000, ymn=1, ymx=100000)
values(test_data) = runif(100, 1, 1000)

#Calculate the parameters for the simulation:
prep_simulation(test_data)

#Run the simulation:
results = run_simulation(test_data, expanded_D, contact_mat, beta)

plot_simulation(results, dim(expanded_D)[1]/4, 4, pretty=T, title="SIR plot")


}
