% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_simulation.R
\name{prep_simulation}
\alias{prep_simulation}
\title{Performs pre-calculations necessary for the simulation of an epidemic}
\usage{
prep_simulation(spatial_data, R0 = 1.8, sigma = 1/2.6, age = TRUE,
  k_params = NA)
}
\arguments{
\item{spatial_data}{The spatial dataset containing the population data.}

\item{R0}{The desired R0 for the epidemic}

\item{sigma}{The desired recovery rate for the epidemic}

\item{age}{Logical. If TRUE, the simulation will assume four distincts age groups in the population (default).
If FALSE, the simulation will assume homogeneous mixing between age groups in the population.}

\item{k_params}{A list containing the parameter values to be used for the kernel calculation. If omitted,
will use the following default values: dist_c=87; alpha=0.95; p=6.6; p2=1.53; aa=35}
}
\value{
Returns two matrix objects and one value for beta.
}
\description{
Creates the age mixing matrix and the spatial kernel using the \code{\link{load_contact_mat}},
             \code{\link{calc_dist_mat}}, \code{\link{calc_dist_kernel}} and \code{\link{calc_beta}} functions.
}
\examples{

#Create a spatial dataset:
test_data = raster(nrow=10, ncol=10, xmn=1, xmx=100000, ymn=1, ymx=100000)
values(test_data) = runif(100, 1, 1000)

prep_simulation(test_data)


}
